This adds the schema and tmcf + csv for CDC COVID-19 Seroprevalence data (https://data.cdc.gov/Laboratory-Surveillance/Nationwide-Commercial-Laboratory-Seroprevalence-Su/d2tw-32xv). It is a large-scale geographic seroprevalence survey that measures the SARS-CoV-2 antibody levels against SARS-CoV-2 in a surveyed population, which is used to estimate the percent of the population that has previously had COVID-19. This data is collected at the US state and territory level. It is collected and reported over non-regular collection period. This import includes the raw count data collected as well as the extrapolated percentage and estimated total cumulative count of the population that has had COVID-19
JSON output: /cns/jv-d/home/datcom/json/cdc/covid19_seroprevalence/report.json

Note: Due to irregular reporting periods report.json includes percent increase and data hole errors due to comparing dates that are not in the consecutive time series. There are also errors of missing schema some of which is added in this import and others which are preexisting schema in the knowledge graph.
PiperOrigin-RevId: 399720781