The Medical Subject Headings (MeSH) thesaurus is a controlled and hierarchically-organized vocabulary produced by the National Library of Medicine (https://www.nlm.nih.gov/mesh/meshhome.html). It is used for indexing, cataloging, and searching of biomedical and health-related information. MeSH includes the subject headings appearing in MEDLINE/PubMed, the NLM Catalog, and other NLM databases.
This cleans up the import of Medical Subject Heading Data from NLM MeSH and formats it as a  tmcf + csv biomedical import. It also cleans up the schema and adds a new datatype MeSHRecord. In addition, it includes mapping of the MeSHRecords to ChemicalCompounds using a mapping file from PubChem (https://pubchem.ncbi.nlm.nih.gov/). It passed the tests from Prashanth's json tool and the internal v3 staging tool.

PiperOrigin-RevId: 475686752