The United States Adopted Names (USAN) Council is co-sponsored by the American Medical Association, the United States Pharmacopeial Convention, Inc., and the American Pharmaceutical Association. The USAN Council’s goal is to provide meaningful designations for compounds, enhancing prescribing and patient safety. It was organized in 1964 to represent the health professions in the selection of appropriate nonproprietary names for drugs.
USAN stems represent common stems for which chemical and/or pharmacologic parameters have been established. These council-approved stems and their definitions are recommended for use in coining new nonproprietary drug names belonging to an established series of related agents. USAN appropriately incorporates this established class stem system. By doing so, similar compounds maintain a common "family" name that provides immediate recognition.

The data is hosted by the American Medical Association and can be downloaded from https://www.ama-assn.org/about/united-states-adopted-names/united-states-adopted-names-approved-stems. It is nonproprietary, which by definition is in the public domain. It is updated annually.

The data is represented in USAdoptedNameStem nodes. The new schema needed to support this new node type is added. This includes adding a new property 'example' to class Thing. This change also involves removes a duplicate entry of class 'Drug' in the schema and resolves conflicts in it's parent class.

This import passed the java lint tool - link to the output report listed below. It also passed the internal v3 staging browser.

PiperOrigin-RevId: 537941454