"The Pharmacogenomics Knowledge Base, PharmGKB, is an interactive tool for researchers investigating how genetic variation affects drug response. The PharmGKB Web site, http://www.pharmgkb.org, displays genotype, molecular, and clinical knowledge integrated into pathway representations and Very Important Pharmacogene (VIP) summaries with links to additional external resources. Users can search and browse the knowledgebase by genes, variants, drugs, diseases, and pathways" Thorn, Klein, & Altman, 2013.
This includes all of the primary data from PharmGKB, which includes summary data on chemicals, drugs, genes, genetic variants, and phenotypes. It also includes the relationships data, which documents associations between two chemicals, chemicals and genes, chemicals and variants, two genes, genes and variants, or two variants.

This import passed the java lint tool - link to the output report listed below. It also passed the internal v3 staging browser.

Note, we have decided to update how we generate the Gene dcids. To ensure linking to existing Gene nodes using the old method a temporary csv + tmcf pair was created to link the Gene nodes used in this import to existing Gene nodes in the graph. This should be removed when the old method for generating dcids is deprecated.

PiperOrigin-RevId: 549662347