This brings in data from NCBI Taxonomy in csv+tmcf format. This includes information from the following files downloaded from the ftp backend: * divisions.dmp * names.dmp * host.dmp * nodes.dmp * categories.dmp
The internal documentation for this import can be found [here](https://docs.google.com/document/d/15AtUnBNcIH2DPtR1uGk62I8ge1Bbq3QQ5rDg-hvwKik/edit?pli=1&resourcekey=0-3KpH9dLoXPnEJ15kDaONPQ&tab=t.0). The code and information for this import is being added to the datacommons/data github, and this addition can be found in PR 1014 (https://github.com/datacommonsorg/data/pull/1014/files).

The schema for the BiologicalTaxonomicDivisionEnum, and BiologicalHostEnum, and BiologicalTaxonomicRankEnum are autogenerated by format_ncbi_taxonomy.py and are added to the biomedical schema as the new file ncbi_taxonomy_enum.mcf

"NCBI Taxonomy \"consists of a curated set of names and classifications for all of the source organisms represented in the International Nucleotide Sequence Database Collaboration (INSDC). The NCBI Taxonomy database contains a list of names that are determined to be nomenclaturally correct or valid (as defined according to the different codes of nomenclature), classified in an approximately phylogenetic hierarchy (depending on the level of knowledge regarding phylogenetic relationships of a given group) as well as a number of names that exist outside the jurisdiction of the codes. That is, it focuses on nomenclature and systematics, rather than documenting the description of taxa."

PiperOrigin-RevId: 646311740